Update case translation
Several important changes since I made this a few years ago. I've 
improved the interpretation of federal court names (things like "S.D. Mississippi, Southern Division" previously broke the translator); 
added support for the yearAsVolume and Archive fields (so far only implemented for WL and NCBC citations, but it would be easy to do more); 
removed some fiddly logic that suppressed certain court information (better practice to do that in the Abbreviations filter now);
added functionality to automatically fetch highlighting from the case and store it (with page citation) as a note on the new item.

I've also commented, linted, and refactored to make it look a bit less awful.